Move PlannerError back
